+++ epi/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+* You can control the maximum egui web canvas size with `App::max_size_points`.
+
 
 ## 0.9.0 - 2021-02-07